+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change: add new features to the prediction model
 - Fix: #7 gracefully handle `NotEnoughError`
 - Change: allow titans to die in the dungeon
+- Fix: upgrade resources
 
 ## `3.0`